Add layers module and move Layer classes
