+++ seacharts/layers/__init__.py
@@ -1,0 +1,1 @@
+from .layer import Layer, Land, Shore, Seabed
