+++ seacharts/layers/layer.py
@@ -1,0 +1,116 @@
+from __future__ import annotations
+
+from abc import ABC
+from dataclasses import dataclass, field
+
+from shapely import geometry as geo
+
+from seacharts.spatial.base import Shape
+
+
+@dataclass
+class Layer(Shape, ABC):
+    @property
+    def _external_labels(self) -> list[str]:
+        raise NotImplementedError
+
+    @property
+    def name(self) -> str:
+        return self.__class__.__name__
+
+    @property
+    def label(self) -> str:
+        return self.name.lower()
+
+    def records_as_geometry(self, records: list[dict]) -> None:
+        if len(records) > 0:
+            self.geometry = self._record_to_geometry(records[0])
+            if isinstance(self.geometry, geo.Polygon):
+                self.geometry = self.as_multi(self.geometry)
+
+    def unify(self, records: list[dict]) -> None:
+        geometries = [self._record_to_geometry(r) for r in records]
+        self.geometry = self.collect(geometries)
+
+
+@dataclass
+class ZeroDepth:
+    depth = 0
+
+
+@dataclass
+class SingleDepth:
+    depth: int
+
+
+@dataclass
+class MultiDepth:
+    @property
+    def depth(self) -> None:
+        raise AttributeError("Multi-depth shapes have no single depth.")
+
+
+@dataclass
+class Locations(Layer, ABC):
+    geometry: geo.MultiPoint = field(default_factory=geo.MultiPoint)
+
+
+@dataclass
+class ZeroDepthLocations(Locations, ZeroDepth, ABC):
+    pass
+
+
+@dataclass
+class SingleDepthLocations(Locations, SingleDepth, ABC):
+    pass
+
+
+@dataclass
+class MultiDepthLocations(Locations, MultiDepth, ABC):
+    pass
+
+
+@dataclass
+class Regions(Layer, ABC):
+    geometry: geo.MultiPolygon = field(default_factory=geo.MultiPolygon)
+
+
+@dataclass
+class ZeroDepthRegions(Regions, ZeroDepth, ABC):
+    pass
+
+
+@dataclass
+class SingleDepthRegions(Regions, SingleDepth, ABC):
+    @property
+    def name(self) -> str:
+        return self.__class__.__name__ + f"{self.depth}m"
+
+
+@dataclass
+class MultiDepthRegions(Regions, MultiDepth, ABC):
+    pass
+
+
+@dataclass
+class Seabed(SingleDepthRegions):
+    color = "seabed"
+    z_order = -300
+    _external_labels = [
+        dict(layer="dybdeareal", depth="minimumsdybde"),
+        dict(layer="grunne", depth="dybde"),
+    ]
+
+
+@dataclass
+class Land(ZeroDepthRegions):
+    color = "land"
+    z_order = -100
+    _external_labels = ["landareal"]
+
+
+@dataclass
+class Shore(ZeroDepthRegions):
+    color = "shore"
+    z_order = -200
+    _external_labels = ["skjer", "torrfall", "landareal", "ikkekartlagtsjomaltomr"]
