+++ seacharts/spatial/__init__.py
@@ -1,3 +1,2 @@
-from .base import Shape
 from .hypsometry import Hydrography, Topography
 from .shapes import Area, Arrow, Circle, Line, Rectangle, Ship
